requirements_full.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0

# ML and embeddings
numpy==1.24.3
sentence-transformers==2.2.2

# Azure services
azure-search-documents==11.4.0
azure-core==1.29.5
azure-identity==1.15.0
openai==1.3.7

# Database
pyodbc==5.0.1
aioodbc==0.4.0

# Additional dependencies
email-validator==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
